requirements.txt
certifi==2019.3.9
correlation-cuda==0.0.0
cycler==0.10.0
joblib==0.11
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
opencv-python==3.4.3.18
Pillow==6.0.0
pyparsing==2.4.0
pypng==0.0.19
python-dateutil==2.8.0
six==1.12.0
torch==1.0.1
